worker/requirements.txt
# Celery and message queue
celery==5.3.6
redis==5.0.1
kombu==5.3.4

# Django (for ORM access to shared models)
Django==5.2.8
psycopg2-binary==2.9.11

# Channels for WebSocket updates
channels==4.2.0
channels-redis==4.2.1
asgiref==3.8.1

# Environment
python-dotenv==1.0.1

# Cryptography for token decryption
cryptography==41.0.7

# HTTP requests for GitHub API
requests==2.31.0

# Git operations (for future scanner)
gitpython==3.1.40

# Future: Security scanners
# bandit==1.7.6
# pip-audit==2.6.1
# safety==2.3.5
